clean up window management
