+++ noteworthy-electron/src/main/global.d.ts
