+++ noteworthy-electron/src/main/index.ts
@@ -18,40 +18,6 @@ import icon from '../../resources/icon.png?asset'
 
 ////////////////////////////////////////////////////////////////////////////////
 
-// function createWindow(): void {
-//   // Create the browser window.
-//   const mainWindow = new BrowserWindow({
-//     width: 900,
-//     height: 670,
-//     show: false,
-//     autoHideMenuBar: true,
-//     ...(process.platform === 'linux' ? { icon } : {}),
-//     webPreferences: {
-//       preload: join(__dirname, '../preload/preload.js'),
-//       sandbox: false
-//     }
-//   })
-
-//   mainWindow.on('ready-to-show', () => {
-//     mainWindow.show()
-//   })
-
-//   mainWindow.webContents.setWindowOpenHandler((details) => {
-//     shell.openExternal(details.url)
-//     return { action: 'deny' }
-//   })
-
-//   // HMR for renderer base on electron-vite cli.
-//   // Load the remote URL for development or the local html file for production.
-//   if (is.dev && process.env['ELECTRON_RENDERER_URL']) {
-//     mainWindow.loadURL(process.env['ELECTRON_RENDERER_URL'])
-//   } else {
-//     mainWindow.loadFile(join(__dirname, '../renderer/index.html'))
-//   }
-// }
-
-////////////////////////////////////////////////////////////////////////////////
-
 async function createAppMenu(app: NoteworthyApp, themeService: ThemeService) {
 	const appMenuTemplate = await makeAppMenuTemplate(app, themeService);
 	const appMenu = Menu.buildFromTemplate(appMenuTemplate);
@@ -97,20 +63,20 @@ app.whenReady().then(() => {
     optimizer.watchWindowShortcuts(window)
   })
 
-  createNoteworthy()
+  createNoteworthy();
 
   // app.on('activate', function () {
   //   // On macOS it's common to re-create a window in the app when the
   //   // dock icon is clicked and there are no other windows open.
   //   if (BrowserWindow.getAllWindows().length === 0) createWindow()
   // })
-})
+});
 
 // Quit when all windows are closed, except on macOS. There, it's common
 // for applications and their menu bar to stay active until the user quits
 // explicitly with Cmd + Q.
 app.on('window-all-closed', () => {
   if (process.platform !== 'darwin') {
-    app.quit()
+    app.quit();
   }
-})
+});
