+++ noteworthy-electron/src/patch.d.ts
@@ -1,11 +1,5 @@
 // type declarations
 
-// declare module NodeJS {
-// 	interface Global {
-// 		isQuitting: boolean;
-// 	}
-// }
-
 interface ProxyConstructor {
 	new <TSource extends object, TTarget extends object>(target: TSource, handler: ProxyHandler<TSource>): TTarget;
 }
